ChhayaAI/ai_agent_service/requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-cloud-spanner>=3.40.0
redis>=5.0.0
